+++ app/hooks/useId.ts
@@ -1,0 +1,9 @@
+import { useMemo } from 'react';
+import { randomString } from '@togglecorp/fujs';
+
+function useId() {
+    const id = useMemo(() => randomString(), []);
+    return id;
+}
+
+export default useId;
