+++ app/views/CountryProfile/IduMap/styles.css
@@ -1,0 +1,4 @@
+.map-container {
+    width: 100%;
+    min-height: 44rem;
+}
